+++ packages/app/src/library/GenerateNominations/Controls/ListControls.tsx
@@ -7,8 +7,8 @@ import { useUnstaking } from 'hooks/useUnstaking'
 import { useState } from 'react'
 import { ButtonMonoInvert, ButtonPrimaryInvert } from 'ui-buttons'
 import { Popover } from 'ui-core/popover'
-import { SelectableWrapper } from '../List'
-import type { ListControlsProps } from './types'
+import { SelectableWrapper } from '../../List'
+import type { ListControlsProps } from '../types'
 
 export const ListControls = ({
   selectHandlers,
